refactor(shell,core): dissolve shellctx into shell and new core package
Cleanup code generation is shell knowledge and moves to
shell/cleanup.go (using the shared Fish constant). Active-chain and
cleanup calculation move to the new internal/core package, which will
host the orchestration engine; its AuthChecker indirection is replaced
by the existing config.AuthChecker interface.
